requirements.txt
shapely~=1.7.1
scikit-image==0.17.2
imgaug==0.4.0
pyclipper~=1.3.0
lmdb~=1.2.1
tqdm~=4.62.3
numpy~=1.19.3
visualdl
python-Levenshtein
opencv-contrib-python==4.4.0.46
lxml~=4.6.3
premailer~=3.10.0
openpyxl~=3.0.9
six~=1.16.0
opencv-python~=4.5.3.56
pillow~=8.4.0
scipy~=1.7.1
requests~=2.26.0
pyyaml~=6.0
setuptools~=58.0.4